Разбивает проект на модули
