+++ js/const.js
@@ -1,0 +1,42 @@
+const ADS_COUNT = 10;
+
+const Coords = {
+  X_MIN: 35.65000,
+  X_MAX: 35.70000,
+  Y_MIN: 139.70000,
+  Y_MAX: 139.80000,
+  PRECISION: 5,
+};
+
+const Offer = {
+  TITLES: ['Title 1', 'Title 2', 'Title 3', 'Title 4', 'Title 5', 'Title 6', 'Title 7', 'Title 8', 'Title 9', 'Title 10'],
+  TYPES: ['palace', 'flat', 'house', 'bungalow'],
+  FEATURES: ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'],
+  CHECKS: ['12:00', '13:00', '14:00'],
+  PHOTOS: [
+    'http://o0.github.io/assets/images/tokyo/hotel1.jpg',
+    'http://o0.github.io/assets/images/tokyo/hotel2.jpg',
+    'http://o0.github.io/assets/images/tokyo/hotel3.jpg',
+    'http://o0.github.io/assets/images/tokyo/hotel4.jpg',
+    'http://o0.github.io/assets/images/tokyo/hotel5.jpg',
+    'http://o0.github.io/assets/images/tokyo/hotel6.jpg',
+    'http://o0.github.io/assets/images/tokyo/hotel7.jpg',
+    'http://o0.github.io/assets/images/tokyo/hotel8.jpg',
+    'http://o0.github.io/assets/images/tokyo/hotel9.jpg',
+    'http://o0.github.io/assets/images/tokyo/hotel10.jpg',
+  ],
+  DESCRIPTION: [
+    'Description 1',
+    'Description 2',
+    'Description 3',
+    'Description 4',
+    'Description 5',
+    'Description 6',
+    'Description 7',
+    'Description 8',
+    'Description 9',
+    'Description 10',
+  ],
+};
+
+export {ADS_COUNT, Coords, Offer};
